lab 2 part 1.2

+++ projects/lpc40xx_freertos/l5_application/gpio_lab.c
@@ -18,13 +18,13 @@ void gpiox__set_as_input(port_pin_s pin_num) {
 }
 
 /// Should alter the hardware registers to set the pin as output
-void gpiox__set_as_output(int pin_num) { LPC_GPIO1->DIR |= (1 << pin_num); }
+void gpiox__set_as_output(uint8_t pin_num) { LPC_GPIO1->DIR |= (1 << pin_num); }
 
 /// Should alter the hardware registers to set the pin as high
-void gpiox__set_high(int pin_num) { LPC_GPIO2->SET = (1 << pin_num); }
+void gpiox__set_high(uint8_t pin_num) { LPC_GPIO2->SET = (1 << pin_num); }
 
 /// Should alter the hardware registers to set the pin as low
-void gpiox__set_low(int pin_num) { LPC_GPIO2->CLR = (1 << pin_num); }
+void gpiox__set_low(uint8_t pin_num) { LPC_GPIO2->CLR = (1 << pin_num); }
 
 /*(void gpiox__set(port_pin_s pin_num, bool high) {
   if (high == 1) {
